implemented two-host architecture

+++ src/Command/StockResult.php
@@ -10,6 +10,6 @@ class StockResult
         public string $action,
         public ?string $sku,
         public ?string $orderId,
-        public ?int $price = null,
+        public ?float $price = null,
     ){}
 }
