+++ src/Contracts/ViewInterface.php
@@ -7,7 +7,7 @@ use Warehouse\Command\StockResult;
 
 interface ViewInterface
 {
-    public function showAction(string $action, ?string $sku, ?string $orderId, ?int $price): void;
+    public function showAction(string $action, ?string $sku, ?string $orderId, ?float $price): void;
     public function showHoldSuccess(string $sku, string $orderId): void;
     public function showConfirmSuccess(string $orderId, string $sku): void;
     public function showNotFound(): void;
